Explicitly mention what `disable_trace` does and the side effects
We used to have something in the README about `disable_trace` hinting
that it was incompatible with any tools that used `trace` functionality
and that for that reason Shopify didn't really use it that much.

It was super-surprising to me when I added Bootsnap to suddenly have
engineer debug workflows affected by it. A better description of the
side effect of the setting might have helped me narrow it down quicker
or prevented me shipping it in the first place with traces disabled.