Only retry reloading from the top level
The underlying load_missing_constant implementation will call itself
recursively (through const_missing). If the constant can't be loaded
both with and without bootsnap disabled, the retries will make this
exponential 😱.

This commit avoids this by setting a thread-local variable inside
load_missing_constant to disable retries. This variable is unset in
require_or_load so that calls to load_missing_constant that happen when
requiring a file are again cached.