Use module_function instead of extend(self)
Kernel.require and Kernel.load are public methods but Kernel#require and
Kernel#load are private.

When we use `extend(self)` ruby defines both instance and class methods
as public in the class that includes this module.

`module_function` defines a public class method but keep the instance
methos as private in the target class.

Closes #233.