expand_path correctly when resolving relative paths:
Without this change, this code:

    autoload(:B, './b')

Resolves to a path:

    $PWD/b # for example, /Users/me/src/foo/bar/app/lib/b

This is missing the .rb extension, which is normally fine because ruby
is happy to resolve it when we pass it to require, etc., but we also
insert that resolved path into our LoadedFeaturesIndex. Later, if we
attempt to purge that file from the LFI, we will generally have the full
path available, and therefore remove a non-existent entry, leaving the
not-fully-expanded one.