Disable iseq cache in Ruby 2.5
There are many bugs in Ruby 2.5 related with iseq cache making it hard
to use tracepoints with it. This means that debuggers and any gem that
uses tracepoint don't play well with bootsnap when used with Ruby 2.5.

Until those issues are fixed in Ruby 2.5 we are going to disabled iseq
cache in the `bootsnap/setup` file.