Update `Bootsnap::LoadPathCache::Store` to better handle concurrent transactions.
In our build system we occasionally see the error
```
Bootsnap::LoadPathCache::Store::NestedTransactionError:
  Bootsnap::LoadPathCache::Store::NestedTransactionError
...
```

However, there is no other call to `Bootsnap::LoadPathCache::Store#transaction` in the callstack. This suggests that we have multiple threads modifying the store at the same time.

Reading `Bootsnap::LoadPathCache::Cache` concurrency is already being accounted for with a `Mutex`. Which means this isn't concurrent calls to the same `Bootsnap::LoadPathCache::Cache`. And looking at `Bootsnap::LoadPathCache.setup` there is a single `Bootsnap::LoadPathCache::Store` that is used by `@load_path_cache` and, if active support in `@autoload_paths_cache`.

Given this `Bootsnap::LoadPathCache::Store` needs to be able to deal with multithreading. This replaces `@in_txn` with a mutex which gives the desired synchronization.

This should not lead to deadlocks with the `Bootsnap::LoadPathCache::Cache` mutexes because the `Bootsnap::LoadPathCache::Store` doesn't make any calls into the `Cache` and the `Cache` already makes all `Bootsnap::LoadPathCache::Store#transaction` calls in a `Mutex#synchronize` block.

This may be related to https://github.com/Shopify/bootsnap/issues/119 since the example stack there also only has `transaction` in it once.