Enable testing using Github Actions
This commit changes the Rakefile to use the minitest rakefile methods to run all of the tests. This is done because the existing `bin/testunit` calls directly to `sh` which is not present on Windows platforms. By making this a ruby/rake call it makes the test suite platform independent. The existing files in `bin` are left in case there is other existing infrastructure that uses them.

This commit also adds a Github Action yaml file that tests this code against the latest versions of windows, ubuntu, and macos. This validates that the test runner change works on all of the existing supported platforms as well as Windows.