Ignore load path caches generated by a different ruby version
Fix: https://github.com/Shopify/bootsnap/issues/384

Until now it was assume that if the `$LOAD_PATH` was identical,
then the content of the paths would be too.

However from one minor ruby version to another, the layout of the
stdlib can change. So if the newer version is installed in the same
place than the previous one, it might cause the load path cache to
be invalid.

So we store `RUBY_DESCRIPTION` in the cache and make sure it matches
before reusing the cache.