Add readonly mode
Fix: https://github.com/Shopify/bootsnap/issues/423

When you know that the cache won't be used again, it avoid
some useless work and IOs.

Typically this might be the case for dockerized applications. You
generate a cache when building the image, but then when you boot the
application any cache update won't be persisted, so thre is no point.